Double checking the LAN interface
Let's back up first and actually check if our Ubuntu VM is using the LAN interface and not the WAN Interface.
As you can see on PfSense, the WAN has it's interface em0 set to 10.0.2.15, while LAN is set to 192.168.1.1
We have to make sure that the internal network is set to LAN, otherwise the process will work in reverse and we won't pick up on any traffic.
To verify this is the case, I had to go to Network setting on the Ubuntu VM, and set adapter aka interface 1 to internal network. This is represented by intnet. This as the name implies is our LAN (Local Area Network)
On the PfSense VM settings, you want to make sure adapter 1 is set to the WAN, and all I did was put a NAT (Network Address Translation) so it converts the external network IP into something our own machine can recognize
You want to make sure on the PFSense Machine, that adapter 2 is set to intnet as well